+++ src/renderer/light/directional_light.rs
@@ -161,7 +161,7 @@ impl Light for DirectionalLight {
         program.use_uniform(&format!("direction{}", i), self.direction.normalize());
     }
 
-    fn id(&self) -> u32 {
+    fn id(&self) -> u8 {
         if self.shadow_texture.is_some() {
             3
         } else {
